+++ R/gldrs.R
@@ -44,34 +44,46 @@ rGLDrs <- function(n, l1, l2, l3, l4){
   qGLDrs(runif(n), l1, l2, l3, l4)
 }
 
-#' @param q quantile value for which the corresponding cumulative probability value should be found.
-#' @param tol numeric tolerance balue for approximating CDF. Default 1e-15
-#' @param maxiter numeric maximum number of iteration
-#' @param n_grid integer size of helper grid to be passed to `make_pgrid`. Default is 50
-#' @param s_grid integer beta shape of helper grid to be passed to `make_pgrid`. Default is 5
+#' @param q vector of quantiles
+#' @param ... used by method
+#' @param lower,upper the `stats::uniroot` lower and upper end points of the interval to be searched. Defaults are 0 and 1, respectively
+#' @param tol the `stats::uniroot` desired accuracy (convergence tolerance). Default value 1e-06
+#' @param silent the `base::try` argument. Default is TRUE
+#' @param trace integer number passed to `stats::uniroot`; if positive, tracing information is produced. Higher values giving more details.
 #' @rdname GLDrs
+#' @importFrom stats uniroot
+#' @include iqf.R
 #' @export
-pGLDrs <- function(q, l1, l2, l3, l4, n_grid=50L, s_grid=5L, tol=1e-15, maxiter=1e3){
+pGLDrs <- iqf(qGLDrs)
 
-  afun <- function(u, q, l1, l2, l3, l4) {q - qGLDrs(u,l1, l2, l3, l4)}
-  p_grd <- sort(c(tol, qpd::make_pgrid(n=n_grid, s=s_grid), 1-tol))
-  q_grd <- qGLDrs(p_grd, l1, l2, l3, l4)
-  idx_lower <- findInterval(q, q_grd, all.inside = TRUE)
-  idx_upper <- idx_lower+1L
-  int_lower <- p_grd[idx_lower]
-  int_upper <- p_grd[idx_upper]
-  ps <- mapply(function(.q, .il, .iu) {
-    tmp_us <- NULL
-    tmp_us <- stats::uniroot(afun, q=.q, l1=l1, l2=l2, l3=l3, l4=l4,
-                             interval=c(.il, .iu), extendInt="no", check.conv=TRUE, tol = tol, maxiter = maxiter)
-    if(is.null(tmp_us)) res <- NA else res <- tmp_us$root
-    #tmp_ps
-  },  q, int_lower, int_upper)
-  ps <- pmin(1, pmax(0, ps))
-
-  ps[!is.finite(ps)] <- NA
-  ps
-}
+# #' @param q quantile value for which the corresponding cumulative probability value should be found.
+# #' @param tol numeric tolerance balue for approximating CDF. Default 1e-15
+# #' @param maxiter numeric maximum number of iteration
+# #' @param n_grid integer size of helper grid to be passed to `make_pgrid`. Default is 50
+# #' @param s_grid integer beta shape of helper grid to be passed to `make_pgrid`. Default is 5
+# #' @rdname GLDrs
+# #' @export
+# pGLDrs <- function(q, l1, l2, l3, l4, n_grid=50L, s_grid=5L, tol=1e-15, maxiter=1e3){
+#
+#   afun <- function(u, q, l1, l2, l3, l4) {q - qGLDrs(u,l1, l2, l3, l4)}
+#   p_grd <- sort(c(tol, qpd::make_pgrid(n=n_grid, s=s_grid), 1-tol))
+#   q_grd <- qGLDrs(p_grd, l1, l2, l3, l4)
+#   idx_lower <- findInterval(q, q_grd, all.inside = TRUE)
+#   idx_upper <- idx_lower+1L
+#   int_lower <- p_grd[idx_lower]
+#   int_upper <- p_grd[idx_upper]
+#   ps <- mapply(function(.q, .il, .iu) {
+#     tmp_us <- NULL
+#     tmp_us <- stats::uniroot(afun, q=.q, l1=l1, l2=l2, l3=l3, l4=l4,
+#                              interval=c(.il, .iu), extendInt="no", check.conv=TRUE, tol = tol, maxiter = maxiter)
+#     if(is.null(tmp_us)) res <- NA else res <- tmp_us$root
+#     #tmp_ps
+#   },  q, int_lower, int_upper)
+#   ps <- pmin(1, pmax(0, ps))
+#
+#   ps[!is.finite(ps)] <- NA
+#   ps
+# }
 
 #' @export
 #' @rdname GLDrs